+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All Notable changes to *jkphl/rdfa-lite-microdata* will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/). This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.0] - 2017-03-03
+### Added
+* Support for HTML5 documents
+* Support for DOMDocument parsing
+
 ## [0.1.0] - 2017-02-20
 
 Initial release